% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtre.R
\name{data_filtre}
\alias{data_filtre}
\title{data_filtre}
\usage{
data_filtre(df, species, sepal_petal, max)
}
\arguments{
\item{df}{dataframe iris a filtrer}

\item{species}{character "setosa", "versicolor" ou  "virginica"}

\item{sepal_petal}{character "Sepal.Length" ou "Petal.Length"}

\item{max}{numeric maximum de "sepal" ou "petal" Length}
}
\value{
dataframe
}
\description{
data_filtre
}
\examples{
data_filtre(iris, "setosa", "Sepal.Length", 6.2)
}
